+++ Bootstrap/Utils/AnalyticsBlocker.h
@@ -6,6 +6,7 @@
 class AnalyticsBlocker
 {
 public:
+	static bool Enabled;
 	static bool ShouldDAB;
 	static bool Initialize();
 	static void Hook();